+++ core/src/test/java/es/urjc/etsii/grafo/util/test/findtypes/DoNotFindMe.java
@@ -1,0 +1,4 @@
+package es.urjc.etsii.grafo.util.test.findtypes;
+
+public class DoNotFindMe {
+}