+++ core/src/test/java/es/urjc/etsii/grafo/util/test/findtypes/FindMe.java
@@ -1,0 +1,7 @@
+package es.urjc.etsii.grafo.util.test.findtypes;
+
+import es.urjc.etsii.grafo.annotations.InheritedComponent;
+
+@InheritedComponent
+public class FindMe {
+}