+++ CHANGELOG.md
@@ -1,13 +1,24 @@
 # Developing
+- Coming soon!
+
+# v0.17
+- (New) Autoconfig module: automatically build and test algorithms, using Irace as the parameter optimizer.
 - (New) StringUtil reverse, longestCommonPrefix, longestCommonSuffix
 - (New) Remove lcp and lcs from instance names
-- (New) Autoconfig: Add Enum types compatibility
 - (New) BitSet implementation compatible with JDK APIs
 - (New) Implement #102: Instance custom properties
 - (Breaking change) maximize as a boolean has been changed in all algorithm components to FMode.MINIMIZE and FMode.MAXIMIZE
 - (Change): Update all dependencies
 - (Change): Improve ScatterSearch internals: refactor code and improve performance, implement missing unit tests
+- (Change): Improve Docker builds: simplify and use multistage
 - (Fix) #107 Autoconfig: Cannot assign integer values to double parameters in algorithm components
+- (Fix) #158: Add TimeControl to SimmulatedAnnealing @isaaclo97
+- (Fix) #163: ArrayUtil sum method truncates double values in some cases @JavierYuste
+- (Fix) #180 Autoconfig: Improve type conversions for algorithm components
+- (Fix) #169 Autoconfig: Remove middleware.sh and improve parallelism 
+- (Fix) #172: Update URL for the project generator
+- (Fix) #130: Remove mandatory implementation of all component types in algorithm components
+- And many more small fixes and improvements
 
 # v0.16
 - (New) Generic ScatterSearch algorithm implementation.